Add files via upload
First crack at a test bench written in Python. The loop variables a and b run from 0 to 15, which can be used directly for unsigned mode, while they are offset by -8 for signed mode in order to cover the range -8 to +7. The dut._log.info that write out test multiplicand, multiplier, and product may be removed.